Update regional workflow based on the test from running: (#553)
* Update regional workflow based on the tests from:
1) RRFS_A v0.6.1 on Cactus
2) summer retro of RRFS determinstic, EnKF DA, ensemble forecast on CONUS 3km on WCOSS2
3) summer retro of EnKF DA on CONUS 3km on Hera

* Remove "export FORT_FMT_NO_WRAP_MARGIN=true" on WCOSS2 model forecast.
This solved the problem that "-DFASTER=ON" hangs the forcast.

+++ scripts/exregional_run_post.sh
@@ -352,8 +352,18 @@ fi
 #   clean forecast netcdf files for saving space
 #-----------------------------------------------------------------------
 #
-rm -f ${dyn_file}
-rm -f ${phy_file}
+if [ ${PREDEF_GRID_NAME} = "RRFS_NA_3km" ]; then
+  indx="00 06 12 18"
+  for i in $indx
+  do
+    if [ "$cyc" == $i ]; then
+      echo "long forecast cycle, keep .nc for bufrsnd" 
+    else
+      rm -f ${dyn_file}
+      rm -f ${phy_file}
+    fi
+  done
+fi
 #
 #-----------------------------------------------------------------------
 #
